requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==5.0.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.115.5
firebase_admin==7.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
pillow==12.1.0
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.3.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
reportlab==4.4.9
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
watchfiles==1.1.1
websockets==16.0
